Add outputAmplitude member/accessors to awg class
In preparation for adding output amplitude specification to the pulse
specification file, we're adding it to the awg class itself.  To maintain
backward compatibility with the class, it is the last member of the
constructor and is optional.

Intended to contain the pk-pk amplitude, in Volts.  (Unchecked) valid values
for the AWG 2040/2041 are 0.020 to 2.0.  Negative values are also invalid.